Trait forest_beacon::Beacon

pub trait Beaconwhere
    Self: Sized,{
    fn verify_entry<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        curr: &'life1 BeaconEntry,
        prev: &'life2 BeaconEntry
    ) -> Pin<Box<dyn Future<Output = Result<bool, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn entry<'life0, 'async_trait>(
        &'life0 self,
        round: u64
    ) -> Pin<Box<dyn Future<Output = Result<BeaconEntry, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn max_beacon_round_for_epoch(
        &self,
        network_version: NetworkVersion,
        fil_epoch: ChainEpoch
    ) -> u64;
}

Trait used as the interface to be able to retrieve bytes from a randomness beacon.

Required Methods

fn verify_entry<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
curr: &'life1 BeaconEntry,
prev: &'life2 BeaconEntry
) -> Pin<Box<dyn Future<Output = Result<bool, Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Verify a new beacon entry against the most recent one before it.

fn entry<'life0, 'async_trait>(
&'life0 self,
round: u64
) -> Pin<Box<dyn Future<Output = Result<BeaconEntry, Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Returns a BeaconEntry given a round. It fetches the BeaconEntry from a Drand node over gRPC In the future, we will cache values, and support streaming.

fn max_beacon_round_for_epoch(
&self,
network_version: NetworkVersion,
fil_epoch: ChainEpoch
) -> u64

Returns the most recent beacon round for the given Filecoin chain epoch.